# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Management API",
        "description": "API for hospital management system",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/tests/": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "list routine tests",
                "description": "List Routine Tests Endpoint.",
                "operationId": "listRoutineTests",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Not authorized"
                    }
                }
            }
        },
        "/api/tests/{id}/": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "read routine test",
                "description": "Read Routine Test Endpoint.",
                "operationId": "readRoutineTest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "test's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Doctor"
                ],
                "summary": "update routine test",
                "description": "Update Routine Test Endpoint.",
                "operationId": "updateRoutineTest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "test's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "breathing_rate": {
                                        "type": "number"
                                    },
                                    "blood_pressure": {
                                        "type": "number"
                                    },
                                    "body_temperature": {
                                        "type": "number"
                                    },
                                    "pulse_rate": {
                                        "type": "number"
                                    },
                                    "medical_notes": {
                                        "type": "string"
                                    },
                                    "prescription": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        },
        "/api/tests/patients/{id}/": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "list patient's routine tests",
                "description": "List Patient's Routine Tests Endpoint.",
                "operationId": "listPatientTests",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "patient's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        },
        "/api/tests/doctors/{id}/": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "list all doctor's routine tests",
                "description": "List Doctor's Routine Tests Endpoint.",
                "operationId": "listDoctorTests",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "departement's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        },
        "/api/tests/me/": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "summary": "list current patient's routine tests",
                "description": "List Current Patient's Tests Endpoint.",
                "operationId": "listCurrentPatientTests",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Not authorized"
                    }
                }
            }
        },
        "/api/tests/me/{id}/": {
            "get": {
                "tags": [
                    "Patient"
                ],
                "summary": "read current patient's routine test",
                "description": "Read Current Patient's Test Endpoint.",
                "operationId": "readCurrentPatientTest",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "test's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        },
        "/api/tests/me/patients/": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "summary": "list routine tests made by the current doctor",
                "description": "List Current Doctor Tests Endpoint.",
                "operationId": "listCurrentDoctorTests",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        },
        "/api/tests/me/patients/{id}/": {
            "get": {
                "tags": [
                    "Doctor"
                ],
                "summary": "read routine test made by current doctor's ",
                "description": "Read Current Doctor Test Endpoint.",
                "operationId": "readCurrentDoctorTests",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "test's id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Object Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}